Create LB_Contest5.cpp
Here we were asked to reverse a sequence in a segmented way. A number K will be given with which we need to reverse each of the k elements from behind. If we reach a point where the the sub sequence cannot be formed into a sub sequence of k elements we can let it as it is.