Create 11_Loop_In_LinkedList.cpp
Things Involved- Checking if the loop is present or not : if the fast and the slow pointers meet inside the while loop which will run until both of them are not equal to NULL. 
Finding the meeting point: return the slow pointer after executing the above process.
Finding the Start of the loop in the linked list: Once you have found the Flyodd's meeting point, start the slow pointer at head and let the other pointer be at the intersection. Increment both of them until they meet each other and they would meet at the starting point of the loop.
And why do they meet exactly here ? Look Out!
Eliminating the Loop: start a temp pointer from the intersect pointer and increment till you find a pointer whose next values is again the intersect and that is where you are going to give a null as its next address and thus by terminating the linked list by removing the loop.