Create 13_FirstUniqueCharInString.cpp
We create a hash map to count all the counts of the characters and also their indexes, use key as char and value as a pair of integers representing count and index.
Initialize answer as size of string and loop through the hash map and see if any of the counts has a value as 1, if true then compare the answer and the current index to find the minimum.

But you can also do it without hashing! Just create a array of constant size of 26 characters. Traverse once through the string, pick the index by using s[i]-'a' and increment the value in that index of the array. After this loop you would have got the array populated in the indexes of the alphabet's positions with their counts. Now traverse again through the string from the beginning to see if it has a frequency of 1 and then return that index, or else return -1 upon reaching the end of the code.